refactor(vm): split vm_data_ops.rs into cohesive submodules (§7-8)
`vm_data_ops.rs` (685 lines) exceeded the 500-line module convention. Split
into three cohesive files, byte-identically (function multiset verified
unchanged; visibility and `self`-dispatch preserved):

- `vm_data_ops.rs` (240): data-structure construction ops — make array (real /
  no-flatten), make hash (from values / from pairs), and Capture creation.
- `vm_data_io_ops.rs` (247): say/note/put/print output ops plus their rendering
  helpers (slip flattening, `needs_method_dispatch` gist/Str routing,
  Rat divide-by-zero check, put/str line collectors).
- `vm_data_push_ops.rs` (206): the fast-path `@arr.push(val)` op.

All three files are now under 500 lines.
