+++ src/vm.rs
@@ -153,7 +153,9 @@ mod vm_comparison_ops;
 mod vm_comparison_order_ops;
 mod vm_control_ops;
 mod vm_core_helpers;
+mod vm_data_io_ops;
 mod vm_data_ops;
+mod vm_data_push_ops;
 mod vm_dispatch_helpers;
 mod vm_env_helpers;
 mod vm_exec_dispatch;
